fix: rename utils.ts to utilities.ts following unicorn/prevent-abbreviations
