+++ src/components/utils/development-tools/TanStackRouterDevelopmentTools.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { isProduction } from "../../../common/utils";
+import { isProduction } from "../../../common/utilities";
 
 export const TanStackRouterDevelopmentTools = isProduction
 	? (): null => null
